improve date handling
